requirements.txt
asgiref== 3.8.1
Django== 4.2.13
djangorestframework==3.15.1
django-environ==0.11.2
psycopg2-binary== 2.9.9
sqlparse== 0.5.0
typing_extensions==4.7.1
django-oauth-toolkit==2.3.0
urllib3==2.2.1
python-dotenv==1.0.1
Pillow
django-cors-headers
